First working version

+++ opendc-compute/opendc-compute-service/src/main/java/org/opendc/compute/service/driver/telemetry/GuestCpuStats.java
@@ -38,6 +38,6 @@ public record GuestCpuStats(
         long idleTime,
         long stealTime,
         long lostTime,
-        double capacity,
-        double usage,
-        double utilization) {}
+        float capacity,
+        float usage,
+        float utilization) {}
