+++ opendc-compute/opendc-compute-telemetry/src/main/kotlin/org/opendc/compute/telemetry/table/HostInfo.kt
@@ -30,6 +30,6 @@ public data class HostInfo(
     val name: String,
     val arch: String,
     val coreCount: Int,
-    val coreSpeed: Double,
+    val coreSpeed: Float,
     val memCapacity: Long,
 )
